refactor(charts): give series formatting one module with read and apply (spec 22 task 5)
BuildSeriesShapeProperties, BuildMarker, BuildSmooth, ReadSeriesFormat
and PatchSeriesFormat become ChartSeriesFormatXml.Read and Apply. The
writer stops appending three elements at three points in the series and
makes one call once the data references are in place; Apply inserts each
child by schema order, which is what lets it serve a fresh series and a
loaded one.

The chart-type defaults do not fold into Apply. An explicit
<c:symbol val="auto"/> on a LineWithMarkers series and a
<c:smooth val="1"/> on an XYScatterSmoothLines one are properties of the
chart type, not of the model: a series element that came out of a file
has already had them applied by whoever wrote it, and conjuring them
again would edit a chart nobody asked to change. They get their own
entry point, ApplyChartTypeDefaults, which only the writer calls.

Two behaviours brought into line, both towards what the writer did:
- a c:marker created for a size or a fill alone now says
  <c:symbol val="auto"/>, so the shape does not fall back to the chart
  style. The writer always wrote that symbol; the patcher did not.
- a c:spPr created for a fill and an outline that both turn out to be
  absent is taken away again, the way an empty c:marker already was.

ChartFormatting now holds nothing but InsertOrdered and the child-order
tables. All eight golden fixtures byte-identical; full suite green.
